rename test files, move test superclass to test_utils
